test: split query editor e2e spec
